refactor: Move GovPay business logic into single hook, away from UI
