+++ apps/editor.planx.uk/src/@planx/components/Pay/Public/providers/types.ts
@@ -1,0 +1,14 @@
+export type PaymentProviderName = "govpay" | "stripe";
+
+export interface PaymentActions {
+  startNewPayment: () => Promise<void>;
+  refetchPayment: () => Promise<void>;
+  resumeExistingPayment: () => Promise<void>;
+  handleSuccess: () => void;
+}
+
+export interface UsePaymentProviderResult {
+  actions: PaymentActions;
+  passportKey: string;
+  hasExistingPayment: boolean;
+}
